+++ ereuse_devicehub/inventory/views.py
@@ -258,14 +258,16 @@ class DeviceCreateView(GenericMixin):
             }
         )
         if form.validate_on_submit():
-            snapshot = form.save(commit=False)
+            form.save(commit=False)
             next_url = url_for('inventory.devicelist')
             if lot_id:
                 next_url = url_for('inventory.lotdevicelist', lot_id=lot_id)
-                lots = self.context['lots']
-                lot = lots.filter(Lot.id == lot_id).one()
-                lot.devices.add(snapshot.device)
-                db.session.add(lot)
+                if form.objs:
+                    lots = self.context['lots']
+                    lot = lots.filter(Lot.id == lot_id).one()
+                    lot.devices = lot.devices.union(form.objs)
+                else:
+                    messages.error('Sorry, the device could not be created')
 
             db.session.commit()
             messages.success('Device "{}" created successfully!'.format(form.type.data))